requirements.txt
pandas>=1.5.0
numpy>=1.21.0
snowflake-connector-python>=2.7.0
python-dotenv>=0.19.0
streamlit>=1.22.0
requests>=2.28.0
pyyaml>=6.0.0
loguru>=0.6.0
pytest>=7.0.0
pytest-cov>=4.0.0 